feat(transactions) handle multiple decimal places refactor(transactions) amount field now is an integer test(transactions) decimal places handling test(transactions) new transcation amount format test(accounts) decimal places parsing on balance test(accounts) new transaction amount format

+++ japb_api/transactions/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from ..accounts.models import Account
 class Transaction(models.Model):
-    amount = models.DecimalField(max_digits=19, decimal_places = 2)
+    amount = models.IntegerField()
     description = models.CharField(max_length=500)
     account = models.ForeignKey(Account, on_delete=models.PROTECT)
     date = models.DateTimeField()
